Sherpa supports Python 3
This removes the comment about Sherpa being restricted to Python 2.7 as it supports Python 3 (we don't test against Python 3.4 but do for >= 3.5).

Since I wanted Sherpa to be mentioned on this page I moved it to the intro section. There's probably better ways to do this but not sure what.